GDI resources will no longer be allocated and deallocated on every TimerProc tick - now instantiated on first tick, and released on WM_CLOSE

+++ ColorPicker/wwinmain.c
@@ -37,5 +37,5 @@ INT APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance
 		DispatchMessageW(&Msg);
 	}
 
-	return Msg.wParam;
+	return (INT) Msg.wParam;
 }
